refactor(maps): extract useRulerTool hook + RulerOverlay from ActiveMapStage
Move the measurement/ruler tool (its state, the activate effect, anchor/waypoint
picking, the live-cursor follow, the reset, and the DOM-resolving measurement
memo) into a useRulerTool hook, and the SVG polyline/endpoints/feet-label render
into a RulerOverlay component. The stage now delegates its ruler-active pointer
branches to the hook. No behavior change; verified by the measurement,
measurement-multipoint, and ruler-color e2e specs (13 tests).
